Add export template dialog #132
* Refactor `ImportDialog` by using an abstract `FileDialog`
* Rename `AGImportFile` to `TemplateFile`
* Rename `ImportType` to `TemplateType`
* Add seperators to `Attack Graphs` menu
