+++ src/Dialogs/AggregationFunctionDialog.ts
@@ -1,7 +1,7 @@
 import { FunctionListDialog } from './FunctionListDialog';
 import { AttackgraphFunction } from '../Model';
 import { AttributeRenderer } from '../AttributeRenderer';
-import { AGImportFile, ImportType } from './ImportFileDialog';
+import { TemplateFile, TemplateType } from './FileDialog';
 
 export class AggregationFunctionListDialog extends FunctionListDialog {
   protected title = mxResources.get('attackGraphs.aggregationFunctions');
@@ -23,21 +23,30 @@ export class AggregationFunctionListDialog extends FunctionListDialog {
     }
   }
 
-  protected override importFileCallback(file: AGImportFile, refresh: () => void): void {
+  protected override importFileCallback(file: TemplateFile, callback: () => void): void {
     this.updateItems(file.aggregation_functions);
-    super.importFileCallback(file, refresh);
+    callback();
   }
 
-  protected override getImportType(): ImportType {
-    return ImportType.AggregationFunctions;
+  protected override getImportType(): TemplateType {
+    return TemplateType.AggregationFunctions;
   }
 
   /**
    * Allows to import a list of function items.
-   * Intended to be used by the template import.
+   * Intended to be used by the template import dialog.
    */
-  static importFunctionItems(ui: Draw.UI, items: AttackgraphFunction[]): void {
+  static override importFunctionItems(ui: Draw.UI, items: AttackgraphFunction[]): void {
     const dlg = new AggregationFunctionListDialog(ui, 0, 0);
     dlg.importFunctionItems(items);
+  }
+
+  /**
+   * Allows to export a list of function items.
+   * Intended to be used by the template export dialog.
+   */
+  static override exportFunctionItems(ui: Draw.UI): AttackgraphFunction[] {
+    const dlg = new AggregationFunctionListDialog(ui, 0, 0);
+    return dlg.getFunctionItems();
   }
 }
