+++ src/Resources.ts
@@ -44,6 +44,7 @@ attackGraphs.openAggregationFunctionsDialog=Aggregation Functions...
 attackGraphs.openComputedAttributesDialog=Computed Attributes...
 attackGraphs.openDefaultAttributesDialog=Default Attributes...
 attackGraphs.openImportTemplateDialog=Import Template...
+attackGraphs.openExportTemplateDialog=Export Template...
 attackGraphs.pickIconForAttribute=Pick an Icon for
 attackGraphs.propertyDefinedTwice=Property defined twice!
 attackGraphs.none=None
@@ -78,8 +79,10 @@ attackGraphs.documentation=Documentation...
 attackGraphs.pageReferenceTooltip=Referenced from
 attackGraphs.workerStatus=Background workers in progress to update the attack graph! Please wait...
 attackgraphs.selectImportItems=Select the items you want to import
+attackgraphs.selectExportItems=Select the items you want to export
 attackgraphs.importFromFile=Import from File
 attackgraphs.importFile=Import file
+attackgraphs.exportFile=Export file
 attackgraphs.alertSelectSingleFile=Please select only a single file!
 attackgraphs.alertFileInvalidFormat=This file has an invalid format! Please select another...
 `);
@@ -113,6 +116,7 @@ attackGraphs.openAggregationFunctionsDialog=Aggregierungsfunktionen...
 attackGraphs.openComputedAttributesDialog=Berechnete Attribute...
 attackGraphs.openDefaultAttributesDialog=Standardattribute...
 attackGraphs.openImportTemplateDialog=Template importieren...
+attackGraphs.openExportTemplateDialog=Template exportieren...
 attackGraphs.pickIconForAttribute=Wähle ein Bild für
 attackGraphs.propertyDefinedTwice=Doppeltes Attribut!
 attackGraphs.none=Keine
@@ -147,8 +151,10 @@ attackGraphs.documentation=Dokumentation...
 attackGraphs.pageReferenceTooltip=Referenziert von
 attackGraphs.workerStatus=Hintergrundprozesse in Arbeit, um den Attack-Graph zu aktualisieren! Bitte warten Sie...
 attackgraphs.selectImportItems=Wählen Sie die Elemente aus, die Sie importieren möchten
+attackgraphs.selectExportItems=Wählen Sie die Elemente aus, die Sie exportieren möchten
 attackgraphs.importFromFile=Aus Datei importieren
 attackgraphs.importFile=Datei importieren
+attackgraphs.exportFile=Datei exportieren
 attackgraphs.alertSelectSingleFile=Bitte nur eine einzelne Datei auswählen!
 attackgraphs.alertFileInvalidFormat=Diese Datei hat ein ungültiges Format! Bitte wählen Sie eine andere aus...
   `);
